prov/efa: make the current reverse AV lock free
Replace the uthash cur_reverse_av (explicit and implicit) with an efa_av_array,
so the lookup in the cq read is lock free.

The array is indexed by a 32-bit key built from the 16-bit AHN and the 16-bit
QPN. The halves are bit interleaved (a Morton code) rather than packed as
ahn << 16 | qpn: interleaving keeps k-bit halves inside a 2k-bit index, so the
realistic case of small AHNs and small QPNs stays in the array's inline region
and neighboring pairs stay neighbors in the index space. Packing by shifting
would instead place consecutive AHNs 64K indexes apart and burn one chunk per
AHN. x86 gets the interleave as two PDEP instructions when the build targets
BMI2; everything else uses the portable shift/mask spread.

To address the whole 32-bit key space, efa_av_array's max_idx ceiling grows to
UINT32_MAX and max_idx widens to uint64_t.

+++ prov/efa/src/efa_av_array.c
@@ -10,7 +10,7 @@ int efa_av_array_init_attr(struct efa_av_array **arr_out,
 			   const struct efa_av_array_attr *attr)
 {
 	struct efa_av_array *arr;
-	unsigned max_idx = attr ? attr->max_idx : 0;
+	uint64_t max_idx = attr ? attr->max_idx : 0;
 	unsigned inline_size = (attr && attr->inline_size) ?
 			       attr->inline_size : EFA_AV_ARRAY_INLINE_SIZE;
 	unsigned chunk_size = (attr && attr->chunk_size) ?
@@ -19,14 +19,14 @@ int efa_av_array_init_attr(struct efa_av_array **arr_out,
 	*arr_out = NULL;
 
 	if (max_idx == 0) {
-		size_t def = (size_t) inline_size * chunk_size - 1;
+		uint64_t def = (uint64_t) inline_size * chunk_size - 1;
 
 		max_idx = def > EFA_AV_ARRAY_MAX_IDX_CEILING ?
-			  EFA_AV_ARRAY_MAX_IDX_CEILING : (unsigned) def;
+			  EFA_AV_ARRAY_MAX_IDX_CEILING : def;
 	}
 	if (max_idx > EFA_AV_ARRAY_MAX_IDX_CEILING) {
-		EFA_WARN(FI_LOG_AV, "efa_av_array max_idx %u out of range\n",
-			 max_idx);
+		EFA_WARN(FI_LOG_AV, "efa_av_array max_idx %" PRIu64
+			 " out of range\n", max_idx);
 		return -FI_EINVAL;
 	}
 
@@ -39,9 +39,10 @@ int efa_av_array_init_attr(struct efa_av_array **arr_out,
 	arr->inline_size = inline_size;
 	arr->chunk_size = chunk_size;
 	if (max_idx >= inline_size) {
-		size_t span = (size_t) max_idx - inline_size + 1;
+		uint64_t span = max_idx - inline_size + 1;
 
-		arr->chunk_table_len = (span + chunk_size - 1) / chunk_size;
+		arr->chunk_table_len = (size_t) ((span + chunk_size - 1) /
+						 chunk_size);
 	}
 
 	*arr_out = arr;
